Harden the dependency supply chain
- Fix all RustSec vulnerabilities: bump minreq to 3.0 and tungstenite to
  0.30 (both on rustls 0.23 / rustls-webpki 0.103, the older ones were
  affected by certificate validation advisories), and update the lock
  file (yanked crates, bytes, crossbeam-epoch, ring, idna, quick-xml...).
- Only lock versions published at least 14 days ago, except rustls
  0.23.45 which fixes RUSTSEC-2026-0285.
- Add deny.toml (crates.io only, no wildcard, permissive licenses,
  advisories) and .cargo/audit.toml. The ignored advisories are for
  unmaintained/unsound crates only pulled by the GUI toolkit (iced 0.12).
- Add a CI workflow: fmt, --locked build/clippy/test, cargo audit and
  cargo deny, with actions pinned by commit hash and read-only token.
- Declare the MIT license with an SPDX expression, give the path
  dependencies a version.

+++ bitbox_manager/src/releases.rs
@@ -30,7 +30,7 @@ const MAX_DOWNLOAD_SIZE: usize = 2 * 1024 * 1024;
 #[derive(Debug)]
 pub enum ReleaseError {
     Http(minreq::Error),
-    HttpStatus(i32, String),
+    HttpStatus(u16, String),
     Json(String),
     /// No release found for this product.
     NotFound(Product),
